Move interpolators into separate package
