+++ application/src/test/java/org/opentripplanner/updater/trip/gtfs/interpolation/NoDataBackwardsEarlinessInterpolatorTest.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.updater.trip.gtfs;
+package org.opentripplanner.updater.trip.gtfs.interpolation;
 
 import static com.google.common.truth.Truth.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
