+++ application/src/test/java/org/opentripplanner/updater/trip/gtfs/updater/http/GtfsRealtimeTripUpdateSourceTest.java
@@ -7,8 +7,8 @@ import java.time.Duration;
 import org.junit.jupiter.api.Test;
 import org.opentripplanner.framework.io.HttpHeaders;
 import org.opentripplanner.test.support.ResourceLoader;
-import org.opentripplanner.updater.trip.gtfs.BackwardsDelayPropagationType;
-import org.opentripplanner.updater.trip.gtfs.ForwardsDelayPropagationType;
+import org.opentripplanner.updater.trip.gtfs.interpolation.BackwardsDelayPropagationType;
+import org.opentripplanner.updater.trip.gtfs.interpolation.ForwardsDelayPropagationType;
 
 public class GtfsRealtimeTripUpdateSourceTest {
